fix: refactor confirmation dialog for roles

+++ lib/src/layer_presentation/pages/organization_pages/organization_list_page/widgets/organizations_action_popup_menu_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:genesis/src/core/extensions/localized_build_context.dart';
 import 'package:genesis/src/layer_domain/entities/organization.dart';
 import 'package:genesis/src/layer_presentation/blocs/organizations_bloc/organizations_bloc.dart';
-import 'package:genesis/src/layer_presentation/pages/organization_pages/organization_list_page/widgets/delete_organizations_dialog.dart';
+import 'package:genesis/src/layer_presentation/shared_widgets/confirmation_dialog.dart';
 import 'package:genesis/src/theming/palette.dart';
 
 class OrganizationsActionPopupMenuButton extends StatelessWidget {
@@ -24,8 +24,8 @@ class OrganizationsActionPopupMenuButton extends StatelessWidget {
         onSelected: (value) {
           final organizations = List.generate(1, (_) => organization);
           final child = switch (value) {
-            _PopupBtnValue.deleteOrganization => DeleteOrganizationsDialog(
-              organizations: organizations,
+            _PopupBtnValue.deleteOrganization => ConfirmationDialog(
+              message: context.$.deleteOrganizationConfirmation(organizations.single.name),
               onDelete: () {
                 context.read<OrganizationsBloc>().add(OrganizationsEvent.deleteOrganizations(organizations));
               },
